cleaning up deploy

+++ deploy.sh
@@ -2,16 +2,29 @@
 
 set -e
 
+DOCKER_IMAGE="timeoff:latest"
 CONTAINER_NAME="alpine_timeoff"
 
-#Check for running container & stop it before starting a new one
-if [[ $(docker inspect -f '{{.State.Running}}' ${CONTAINER_NAME}) = "true" ]]; then
-    sudo docker stop alpine_timeoff
+
+# Check for existing image
+if [[ "$(sudo docker images -q $DOCKER_IMAGE 2> /dev/null)" == "" ]]; then
+            echo 'Building Image:'
+            sudo docker build --rm -t timeoff:latest --no-cache .
+                exit 1
 fi
 
-echo "Building Image"
-sudo docker build --rm -t timeoff:latest --no-cache .
-echo "Running Docker Container using Docker Image name: timeoff:latest"
-sudo docker run -t -i -d -p 3000:3000 --name alpine_timeoff timeoff
+#Check for running container & stop it before starting a new one
+if [ "$(sudo docker ps -aq -f status=running -f name=${CONTAINER_NAME})" ]; then
+        echo "Container already exist, cleaning up"
+        sudo docker stop $(sudo docker ps -aqf "name=${CONTAINER_NAME}")
 
-sudo docker ps -a
+        if [ "$(sudo docker ps -aq -f status=exited -f name=${CONTAINER_NAME})" ]; then
+                echo "Removing Existing Container"
+                sudo docker rm $(sudo docker ps -aqf "name=${CONTAINER_NAME}")
+
+        fi
+
+        echo "Starting App using Docker Image name: $CONTAINER_NAME"
+        sudo docker run -t -i -d -p 3000:3000 --name alpine_timeoff timeoff
+        sudo docker ps -a
+fi
